detect labels fix
