+++ pigletprep/src/app/videotomcq/DetectLabels.tsx
@@ -1,0 +1,120 @@
+"use client";
+
+import { useRef, useState } from "react";
+
+interface Label {
+  Name: string;
+  Confidence: number;
+}
+
+interface DetectLabelsProps {
+  videoSrc: string;
+}
+
+const DetectLabels: React.FC<DetectLabelsProps> = ({ videoSrc }) => {
+  const videoRef = useRef<HTMLVideoElement | null>(null);
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+  const [imageData, setImageData] = useState<string | null>(null);
+  const [labels, setLabels] = useState<Label[] | null>(null);
+  const [error, setError] = useState<string | null>(null);
+  const [loading, setLoading] = useState<boolean>(false);
+  const [mcq, setMcq] = useState<string | null>(null);
+
+  const captureScreenshot = () => {
+    const video = videoRef.current;
+    const canvas = canvasRef.current;
+
+    if (!video || !canvas) return;
+
+    const ctx = canvas.getContext("2d");
+    if (!ctx) return;
+
+    canvas.width = video.videoWidth;
+    canvas.height = video.videoHeight;
+    ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
+
+    const dataUrl = canvas.toDataURL("image/png");
+    setImageData(dataUrl);
+
+    sendImageToGPT(dataUrl);
+  };
+
+  const sendImageToGPT = async (base64Image: string) => {
+    try {
+      const base64String = base64Image.split(",")[1];
+
+      const response = await fetch("/api/analyzeImage", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ imageBuffer: base64String }),
+      });
+
+      const data = await response.json();
+
+      if (response.ok) {
+        setMcq(data.mcq);
+        setLabels(data.labels || []);
+      } else {
+        setError(data.error || "Error generating MCQ");
+      }
+    } catch (err) {
+      setError((err as Error).message || "Error processing image");
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <div style={{ textAlign: "center" }}>
+      <video ref={videoRef} width="640" height="360" controls crossOrigin="anonymous">
+        <source src={videoSrc} type="video/mp4" />
+      </video>
+
+      <button onClick={captureScreenshot} disabled={loading} style={{ display: "block", margin: "10px auto", color: "black" }}>
+        {loading ? "Processing..." : "Get Multiple Choice Question"}
+      </button>
+
+      <canvas ref={canvasRef} style={{ display: "none" }}></canvas>
+
+      {imageData && (
+        <div
+          style={{
+            position: "fixed",
+            bottom: "20px",
+            right: "20px",
+            background: "white",
+            padding: "10px",
+            borderRadius: "8px",
+            boxShadow: "0px 4px 6px rgba(0,0,0,0.1)",
+          }}
+        >
+          <img src={imageData} alt="Screenshot" style={{ width: "150px", height: "auto", borderRadius: "5px" }} />
+        </div>
+      )}
+
+      {error && <p style={{ color: "red" }}>{error}</p>}
+
+      {labels && (
+        <div>
+          <h2>Detected Labels:</h2>
+          <ul>
+            {labels.map((label, index) => (
+              <li key={index}>
+                {label.Name} - Confidence: {label.Confidence.toFixed(2)}%
+              </li>
+            ))}
+          </ul>
+        </div>
+      )}
+
+      {mcq && (
+        <div>
+          <h2 style={{ color: "black" }}>Generated Multiple-Choice Question:</h2>
+          <p style={{ color: "black" }}>{mcq}</p>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default DetectLabels;
